# server/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Versus Api",
    "version": "1.0.0",
    "description": "Documentação do sistema de torneios Versus"
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "token"
      }
    }
  },
  "security": [
    {
      "cookieAuth": []
    }
  ],
  "paths": {
...
    "/event": {
      "get": {
        "summary": "Lista todos os eventos",
        "tags": [
          "Eventos"
        ],
        "responses": {
          "200": {
            "description": "Lista de eventos"
          }
        }
      },
      "post": {
        "summary": "Cria um novo evento",
        "tags": [
          "Eventos"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "example": {
                  "name": "Campeonato de Verão",
                  "startDate": "2025-07-01T10:00:00.000Z",
                  "maxPlayers": 8,
                  "multiplayer": true
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Evento criado com sucesso"
          },
          "400": {
            "description": "Erro ao criar evento"
          }
        }
      }
    },
    "/event/{id}": {
      "get": {
        "summary": "Retorna um evento pelo ID",
        "tags": [
          "Eventos"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Evento encontrado"
          },
          "404": {
            "description": "Evento não encontrado"
          }
        }
      },
      "put": {
        "summary": "Atualiza um evento",
        "tags": [
          "Eventos"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "example": {
                  "name": "Campeonato Atualizado",
                  "startDate": "2025-07-05T12:00:00.000Z",
                  "maxPlayers": 16
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Evento atualizado"
          },
          "400": {
            "description": "Erro na atualização"
          }
        }
      },
      "delete": {
        "summary": "Remove um evento",
        "tags": [
          "Eventos"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Evento deletado"
          },
          "400": {
            "description": "Erro na deleção"
          }
        }
      }
    },
    "/event/{id}/inscribe": {
      "post": {
        "summary": "Inscreve o usuário logado no evento",
        "tags": [
          "Eventos"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "example": {
                  "teamId": 3
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Inscrição realizada"
          },
          "400": {
            "description": "Erro na inscrição"
          }
        }
      }
    },
    "/event/{id}/unsubscribe": {
      "post": {
        "summary": "Remove a inscrição do usuário (ou de outro se for admin/organizador)",
        "tags": [
          "Eventos"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "example": {
                  "userId": 7
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Inscrição removida"
          },
          "400": {
            "description": "Erro ao remover inscrição"
          }
        }
      }
    },
    "/event/{id}/start": {
      "post": {
        "summary": "Inicia o evento e gera os confrontos",
        "tags": [
          "Eventos"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Confrontos gerados"
          },
          "400": {
            "description": "Erro ao iniciar evento"
          }
        }
      }
    },
    "/event/{id}/inscriptions": {
      "get": {
        "summary": "Lista todas as inscrições de um evento",
        "tags": [
          "Eventos"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID do evento"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de inscrições do evento",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Erro ao buscar inscrições"
          }
        }
      }
    },
    "/event/inscriptions/me": {
      "get": {
        "summary": "Lista as inscrições do usuário logado",
        "tags": [
          "Eventos"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de inscrições do usuário",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Erro ao buscar inscrições do usuário"
          }
        }
      }
    },
...
  },
...
}